refactor: split custom-downloader into modular files
